Sorting and Big O

+++ 0-bubble_sort.c
@@ -7,7 +7,7 @@
 */
 void bubble_sort(int *array, size_t size)
 {
-	int i, j, temp;
+	size_t i, j, temp;
 
 	if (array == NULL || !size)
 		return;
